refactor(snowflake): localize the pyarrow json type to snowflake
