+++ ibis/backends/snowflake/converter.py
@@ -1,16 +1,52 @@
 from __future__ import annotations
 
 import datetime
+import json
 from typing import TYPE_CHECKING
 
+import pyarrow as pa
+
 from ibis.formats.pandas import PandasData
-from ibis.formats.pyarrow import PYARROW_JSON_TYPE, PyArrowData
+from ibis.formats.pyarrow import PyArrowData
 
 if TYPE_CHECKING:
-    import pyarrow as pa
-
     import ibis.expr.datatypes as dt
     from ibis.expr.schema import Schema
+
+
+class JSONScalar(pa.ExtensionScalar):
+    def as_py(self):
+        value = self.value
+        if value is None:
+            return value
+        else:
+            return json.loads(value.as_py())
+
+
+class JSONArray(pa.ExtensionArray):
+    pass
+
+
+class JSONType(pa.ExtensionType):
+    def __init__(self):
+        super().__init__(pa.string(), "ibis.json")
+
+    def __arrow_ext_serialize__(self):
+        return b""
+
+    @classmethod
+    def __arrow_ext_deserialize__(cls, storage_type, serialized):
+        return cls()
+
+    def __arrow_ext_class__(self):
+        return JSONArray
+
+    def __arrow_ext_scalar_class__(self):
+        return JSONScalar
+
+
+PYARROW_JSON_TYPE = JSONType()
+pa.register_extension_type(PYARROW_JSON_TYPE)
 
 
 class SnowflakePandasData(PandasData):
@@ -50,16 +86,12 @@ class SnowflakePandasData(PandasData):
 class SnowflakePyArrowData(PyArrowData):
     @classmethod
     def convert_table(cls, table: pa.Table, schema: Schema) -> pa.Table:
-        import pyarrow as pa
-
         columns = [cls.convert_column(table[name], typ) for name, typ in schema.items()]
         return pa.Table.from_arrays(columns, names=schema.names)
 
     @classmethod
     def convert_column(cls, column: pa.Array, dtype: dt.DataType) -> pa.Array:
         if dtype.is_json() or dtype.is_array() or dtype.is_map() or dtype.is_struct():
-            import pyarrow as pa
-
             if isinstance(column, pa.ChunkedArray):
                 column = column.combine_chunks()
