+++ ibis/formats/pyarrow.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import json
 from typing import TYPE_CHECKING, Any
 
 import pyarrow as pa
@@ -12,37 +11,6 @@ from ibis.formats import DataMapper, SchemaMapper, TableProxy, TypeMapper
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
-
-
-class JSONScalar(pa.ExtensionScalar):
-    def as_py(self):
-        value = self.value
-        if value is None:
-            return value
-        else:
-            return json.loads(value.as_py())
-
-
-class JSONArray(pa.ExtensionArray):
-    pass
-
-
-class JSONType(pa.ExtensionType):
-    def __init__(self):
-        super().__init__(pa.string(), "ibis.json")
-
-    def __arrow_ext_serialize__(self):
-        return b""
-
-    @classmethod
-    def __arrow_ext_deserialize__(cls, storage_type, serialized):
-        return cls()
-
-    def __arrow_ext_class__(self):
-        return JSONArray
-
-    def __arrow_ext_scalar_class__(self):
-        return JSONScalar
 
 
 _from_pyarrow_types = {
@@ -91,6 +59,7 @@ _to_pyarrow_types = {
     dt.MACADDR: pa.string(),
     dt.INET: pa.string(),
     dt.UUID: pa.string(),
+    dt.JSON: pa.string(),
 }
 
 
@@ -128,8 +97,6 @@ class PyArrowType(TypeMapper):
             key_dtype = cls.to_ibis(typ.key_type, typ.key_field.nullable)
             value_dtype = cls.to_ibis(typ.item_type, typ.item_field.nullable)
             return dt.Map(key_dtype, value_dtype, nullable=nullable)
-        elif isinstance(typ, JSONType):
-            return dt.JSON()
         elif pa.types.is_dictionary(typ):
             return cls.to_ibis(typ.value_type)
         else:
@@ -191,8 +158,6 @@ class PyArrowType(TypeMapper):
                 nullable=dtype.value_type.nullable,
             )
             return pa.map_(key_field, value_field, keys_sorted=False)
-        elif dtype.is_json():
-            return PYARROW_JSON_TYPE
         elif dtype.is_geospatial():
             return pa.binary()
         else:
@@ -303,7 +268,3 @@ class PyArrowTableProxy(TableProxy[pa.Table]):
 
     def to_pyarrow(self, schema: Schema) -> pa.Table:
         return self.obj
-
-
-PYARROW_JSON_TYPE = JSONType()
-pa.register_extension_type(PYARROW_JSON_TYPE)
